customize gatsby-remark-external-links feature
