+++ src/components/Landing/index.js
@@ -11,7 +11,7 @@ import TestimonialsCarrousel from "../Testimonials";
 import With4Geeks from "../With4Geeks";
 // import WhyPython from '../WhyPython';
 import AlumniProjects from "../AlumniProjects";
-import { SuccessStories } from "../../templates/success-stories";
+import SuccessStories  from "../SuccessStories";
 import GeeksVsOthers from "../GeeksVsOthers";
 import GeeksInfo from "../GeeksInfo";
 import ProgramDetails from "../ProgramDetails";
@@ -902,14 +902,14 @@ export const landingSections = {
   testimonials_new: ({ session, data, pageContext, yml, index }) => (
     <Div
       id="testimonials_new"
-      key={index}
+      key={`${index}-testimonials_new`}
       flexDirection="column"
       margin="30px 0 0 0"
       // margin_tablet="100px"
       m_sm="0"
       p_xs="0"
     >
-      <SuccessStories pageContext={pageContext} data={data} yml={yml} />
+      <SuccessStories lang={pageContext.lang} filterIndexes={yml.filter_indexes} />
     </Div>
   ),
