+++ src/session.js
@@ -67,7 +67,6 @@ export default ({ children }) => {
   const [session, setSession] = useState(defaultSession);
   //get ip address
   useEffect(() => {
-    console.log("--- Start of useEffect in Session component  ---");
     const urlParams = new URLSearchParams(window.location.search);
     const getReferral = () => {
       let alias = ["referral_code", "ref", "referral_key", "referral"];
